requirements.txt
django
django-tastypie
MySQL-python
South
PyYaml